requirements.txt
# ------------------------------
# CORE DJANGO + DRF
# ------------------------------
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-rest-passwordreset>=1.3.0
django-filter>=24.2
whitenoise>=6.6
djangorestframework-simplejwt>=5.3.0


# ------------------------------
# DATABASE
# ------------------------------
psycopg2-binary>=2.9

# ------------------------------
# CELERY + REDIS
# ------------------------------
celery>=5.3
redis>=5.0

# ------------------------------
# SUPABASE
# ------------------------------
supabase>=2.3.4
httpx>=0.27

# ------------------------------
# OPENAI + LANGCHAIN
# ------------------------------
openai>=1.51
langchain>=0.2
langchain-openai>=0.1
langchain-community>=0.2
langgraph>=0.2

# ------------------------------
# DOCUMENT / OCR PIPELINE
# ------------------------------
pypdfium2==4.30.0
python-docx==1.1.2
paddleocr==2.7.0.3
paddlepaddle>=3.1.0
opencv-python==4.6.0.66
numpy==1.26.4
pillow==10.4.0
shapely==2.0.4
regex==2024.5.15

# ------------------------------
# UTILS
# ------------------------------
requests>=2.32
python-slugify>=8.0
python-dotenv>=1.0
pydantic>=2.6
tiktoken>=0.7

# ------------------------------
# SERVER
# ------------------------------
gunicorn>=21.0
